@@ -16,14 +16,11 @@
 
     "require": {
         "php":                   ">=5.3.1",
-        "symfony/css-selector":  "^2.7|^3.0|^4.0|^5.0"
+        "symfony/css-selector":  "^2.7|^3.0|^4.0"
     },
 
     "require-dev": {
-        "symfony/debug": "^2.7|^3.0|^4.0",
-        "symfony/phpunit-bridge": "^3.4.38 || ^5.0.5",
-        "phpunit/phpunit": "^4.8.36 || ^5.7.27 || ^6.5.14 || ^7.5.20"
-
+        "symfony/phpunit-bridge": "^4.2"
     },
 
     "suggest": {
@@ -48,7 +45,7 @@
 
     "extra": {
         "branch-alias": {
-            "dev-master": "1.8.x-dev"
+            "dev-master": "1.7.x-dev"
         }
     }
 }
