@@ -30,41 +30,35 @@
     }
   },
   "require": {
-    "php": ">=5.6.0",
+    "php": ">=7.1.3",
     "ext-dom": "*",
-    "chi-teck/drupal-code-generator": "^1.28.1",
+    "chi-teck/drupal-code-generator": "^1.30.5",
     "composer/semver": "^1.4",
-    "consolidation/annotated-command": "^2.12",
     "consolidation/config": "^1.2",
     "consolidation/filter-via-dot-access-data": "^1",
-    "consolidation/output-formatters": "^3.3.1",
-    "consolidation/robo": "^1.4.6",
+    "consolidation/robo": "^1 || ^2",
     "consolidation/site-alias": "^3.0.0@stable",
-    "consolidation/site-process": "^2.0.1",
+    "consolidation/site-process": "^2.1 || ^4",
     "grasmash/yaml-expander": "^1.1.1",
     "league/container": "~2",
     "psr/log": "~1.0",
     "psy/psysh": "~0.6",
-    "symfony/console": "^3.4",
-    "symfony/event-dispatcher": "^3.4",
+    "symfony/event-dispatcher": "^3.4 || ^4.0",
     "symfony/finder": "^3.4 || ^4.0",
-    "symfony/process": "^3.4",
-    "symfony/var-dumper": "^3.4 || ^4.0",
-    "symfony/yaml": "^3.4",
-    "webflo/drupal-finder": "^1.1",
+    "symfony/var-dumper": "^3.4 || ^4.0 || ^5.0",
+    "symfony/yaml": "^3.4 || ^4.0",
+    "webflo/drupal-finder": "^1.2",
     "webmozart/path-util": "^2.1.0"
   },
   "require-dev": {
-    "lox/xhprof": "dev-master",
-    "g1a/composer-test-scenarios": "^3",
-    "phpunit/phpunit": "^4.8.36 || ^6.1",
-    "squizlabs/php_codesniffer": "^2.7",
-    "composer/installers": "^1.2",
+    "composer/installers": "^1.7",
     "cweagans/composer-patches": "~1.0",
-    "webflo/drupal-core-strict": "8.6.x-dev",
-    "drupal/empty_theme": "1.0",
-    "drupal/devel": "^2",
+    "drupal/core-recommended": "^8.8",
     "drupal/alinks": "1.0.0",
+    "g1a/composer-test-scenarios": "^3",
+    "lox/xhprof": "dev-master",
+    "phpunit/phpunit": "^4.8.36 || ^6.1",
+    "squizlabs/php_codesniffer": "^2.7 || ^3",
     "vlucas/phpdotenv": "^2.4"
   },
   "autoload": {
@@ -85,7 +79,7 @@
     "sort-packages": true,
     "process-timeout": 9600,
     "platform": {
-      "php": "7.0.11"
+      "php": "7.1.3"
     }
   },
   "scripts": {
@@ -121,17 +115,8 @@
       "sut/themes/contrib/{$name}": ["type:drupal-theme"],
       "sut/drush/contrib/{$name}": ["type:drupal-drush"]
     },
-    "scenarios": {
-      "php5": {
-        "config": {
-          "platform": {
-            "php": "5.6.38"
-          }
-        }
-      }
-    },
     "branch-alias": {
-        "dev-master": "9.x-dev"
+        "dev-master": "10.x-dev"
     }
   }
 }
