@@ -13,18 +13,16 @@
         }
     ],
     "require": {
-        "php": ">=5.4.0",
+        "php": "^8.0 || ^7.0 || ^5.5.9",
         "ext-json": "*",
         "ext-tokenizer": "*",
-        "symfony/console": "~2.3.10|^2.4.2|~3.0|~4.0",
-        "symfony/var-dumper": "~2.7|~3.0|~4.0",
-        "nikic/php-parser": "~1.3|~2.0|~3.0|~4.0",
-        "dnoegel/php-xdg-base-dir": "0.1",
-        "jakub-onderka/php-console-highlighter": "0.3.*|0.4.*"
+        "symfony/console": "~5.0|~4.0|~3.0|^2.4.2|~2.3.10",
+        "symfony/var-dumper": "~5.0|~4.0|~3.0|~2.7",
+        "nikic/php-parser": "~4.0|~3.0|~2.0|~1.3",
+        "dnoegel/php-xdg-base-dir": "0.1.*"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.8.35|~5.0|~6.0|~7.0",
-        "hoa/console": "~2.15|~3.16",
+        "hoa/console": "3.17.*",
         "bamarni/composer-bin-plugin": "^1.2"
     },
     "suggest": {
@@ -48,7 +46,7 @@
     "bin": ["bin/psysh"],
     "extra": {
         "branch-alias": {
-            "dev-develop": "0.9.x-dev"
+            "dev-master": "0.10.x-dev"
         }
     }
 }
