@@ -3,7 +3,7 @@
 namespace Drupal\crop\Form;
 
 use Drupal\Core\Entity\EntityConfirmFormBase;
-use Drupal\Core\Entity\Query\QueryFactory;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\StringTranslation\TranslationInterface;
 use Drupal\Core\Url;
@@ -15,13 +15,6 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
 class CropTypeDeleteForm extends EntityConfirmFormBase {
 
   /**
-   * The query factory to create entity queries.
-   *
-   * @var \Drupal\Core\Entity\Query\QueryFactory
-   */
-  protected $queryFactory;
-
-  /**
    * String translation manager.
    *
    * @var \Drupal\Core\StringTranslation\TranslationInterface
@@ -31,13 +24,10 @@ class CropTypeDeleteForm extends EntityConfirmFormBase {
   /**
    * Constructs a new CropTypeDeleteForm object.
    *
-   * @param \Drupal\Core\Entity\Query\QueryFactory $query_factory
-   *   The entity query object.
    * @param \Drupal\Core\StringTranslation\TranslationInterface $string_translation
    *   The string translation manager.
    */
-  public function __construct(QueryFactory $query_factory, TranslationInterface $string_translation) {
-    $this->queryFactory = $query_factory;
+  public function __construct(TranslationInterface $string_translation) {
     $this->translation = $string_translation;
   }
 
@@ -46,7 +36,6 @@ class CropTypeDeleteForm extends EntityConfirmFormBase {
    */
   public static function create(ContainerInterface $container) {
     return new static(
-      $container->get('entity.query'),
       $container->get('string_translation')
     );
   }
@@ -76,7 +65,7 @@ class CropTypeDeleteForm extends EntityConfirmFormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state) {
-    $count = $this->queryFactory->get('crop')
+    $count = $this->entityTypeManager->getStorage('crop')->getQuery()
       ->condition('type', $this->entity->id())
       ->count()
       ->execute();
@@ -99,7 +88,7 @@ class CropTypeDeleteForm extends EntityConfirmFormBase {
   public function submitForm(array &$form, FormStateInterface $form_state) {
     $this->entity->delete();
     $t_args = ['%name' => $this->entity->label()];
-    drupal_set_message($this->t('The crop type %name has been deleted.', $t_args));
+    $this->messenger()->addMessage($this->t('The crop type %name has been deleted.', $t_args));
     $this->logger('crop')->notice('Deleted crop type %name.', $t_args);
 
     $form_state->setRedirectUrl($this->getCancelUrl());
