@@ -113,7 +113,7 @@ abstract class CropUnitTestBase extends KernelTestBase {
    *   File object.
    */
   protected function getTestFile() {
-    file_unmanaged_copy(drupal_get_path('module', 'crop') . '/tests/files/sarajevo.png', PublicStream::basePath());
+    \Drupal::service('file_system')->copy(drupal_get_path('module', 'crop') . '/tests/files/sarajevo.png', PublicStream::basePath());
     return $this->fileStorage->create([
       'uri' => 'public://sarajevo.png',
       'status' => FILE_STATUS_PERMANENT,
