@@ -7,7 +7,7 @@ use Drupal\Core\Plugin\Context\Context;
 use Drupal\Core\Plugin\Context\ContextDefinitionInterface;
 
 /**
- * @todo.
+ * Defines context that loads entity on demand.
  */
 class EntityLazyLoadContext extends Context {
 
