@@ -2,11 +2,11 @@
 
 namespace Drupal\ctools\Controller;
 
-use Drupal\Core\Controller\ControllerResolverInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Form\FormBuilderInterface;
 use Drupal\Core\Routing\RouteMatchInterface;
 use Drupal\ctools\Wizard\WizardFactoryInterface;
+use Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface;
 
 /**
  * Wrapping controller for wizard forms that serve as the main page body.
@@ -21,8 +21,8 @@ class WizardEntityFormController extends WizardFormController {
   protected $entityTypeManager;
 
   /**
-   * @param \Drupal\Core\Controller\ControllerResolverInterface $controller_resolver
-   *   The controller resolver.
+   * @param \Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface $argument_resolver
+   *   The argument resolver.
    * @param \Drupal\Core\Form\FormBuilderInterface $form_builder
    *   The form builder.
    * @param \Drupal\ctools\Wizard\WizardFactoryInterface $wizard_factory
@@ -30,8 +30,8 @@ class WizardEntityFormController extends WizardFormController {
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
    *   The entity type manager.
    */
-  public function __construct(ControllerResolverInterface $controller_resolver, FormBuilderInterface $form_builder, WizardFactoryInterface $wizard_factory, EntityTypeManagerInterface $entity_type_manager) {
-    parent::__construct($controller_resolver, $form_builder, $wizard_factory);
+  public function __construct(ArgumentResolverInterface $argument_resolver, FormBuilderInterface $form_builder, WizardFactoryInterface $wizard_factory, EntityTypeManagerInterface $entity_type_manager) {
+    parent::__construct($argument_resolver, $form_builder, $wizard_factory);
     $this->entityTypeManager = $entity_type_manager;
   }
 
