@@ -2,12 +2,12 @@
 
 namespace Drupal\ctools\Controller;
 
-use Drupal\Core\Controller\ControllerResolverInterface;
 use Drupal\Core\Controller\FormController;
 use Drupal\Core\Form\FormBuilderInterface;
 use Drupal\Core\Routing\RouteMatchInterface;
 use Drupal\ctools\Wizard\WizardFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface;
 
 /**
  * Wrapping controller for wizard forms that serve as the main page body.
@@ -36,15 +36,15 @@ class WizardFormController extends FormController {
   protected $dispatcher;
 
   /**
-   * @param \Drupal\Core\Controller\ControllerResolverInterface $controller_resolver
-   *   The controller resolver.
+   * @param \Symfony\Component\HttpKernel\Controller\ArgumentResolverInterface $argument_resolver
+   *   The argument resolver.
    * @param \Drupal\Core\Form\FormBuilderInterface $form_builder
    *   The form builder.
    * @param \Drupal\ctools\Wizard\WizardFactoryInterface $wizard_factory
    *   The wizard factory.
    */
-  public function __construct(ControllerResolverInterface $controller_resolver, FormBuilderInterface $form_builder, WizardFactoryInterface $wizard_factory) {
-    parent::__construct($controller_resolver, $form_builder);
+  public function __construct(ArgumentResolverInterface $argument_resolver, FormBuilderInterface $form_builder, WizardFactoryInterface $wizard_factory) {
+    parent::__construct($argument_resolver, $form_builder);
     $this->wizardFactory = $wizard_factory;
   }
 
