@@ -5,16 +5,17 @@ namespace Drupal\ctools\Form;
 use Drupal\Core\Ajax\AjaxResponse;
 use Drupal\Core\Ajax\CloseModalDialogCommand;
 use Drupal\Core\Ajax\RedirectCommand;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Entity\Plugin\DataType\EntityAdapter;
 use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Plugin\Context\Context;
 use Drupal\Core\Plugin\Context\ContextDefinition;
 use Drupal\Core\Plugin\Context\ContextInterface;
-use Drupal\Core\Url;
+use Drupal\Core\Plugin\Context\EntityContextDefinition;
 use Drupal\Core\TempStore\SharedTempStoreFactory;
+use Drupal\Core\Url;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Drupal\Core\Entity\EntityTypeManagerInterface;
 
 abstract class ContextConfigure extends FormBase {
 
@@ -77,7 +78,12 @@ abstract class ContextConfigure extends FormBase {
       $edit = TRUE;
     }
     else {
-      $context_definition = new ContextDefinition($context_id);
+      if (strpos($context_id, 'entity:') === 0) {
+        $context_definition = new EntityContextDefinition($context_id);
+      }
+      else {
+        $context_definition = new ContextDefinition($context_id);
+      }
       $context = new Context($context_definition);
       $machine_name = '';
     }
@@ -165,7 +171,12 @@ abstract class ContextConfigure extends FormBase {
     $contexts = $this->getContexts($cached_values);
     if ($form_state->getValue('machine_name') != $form_state->getValue('context_id')) {
       $data_type = $form_state->getValue('context_id');
-      $context_definition = new ContextDefinition($data_type, $form_state->getValue('label'), TRUE, FALSE, $form_state->getValue('description'));
+      if (strpos($data_type, 'entity:') === 0) {
+        $context_definition = new EntityContextDefinition($data_type, $form_state->getValue('label'), TRUE, FALSE, $form_state->getValue('description'));
+      }
+      else {
+        $context_definition = new ContextDefinition($data_type, $form_state->getValue('label'), TRUE, FALSE, $form_state->getValue('description'));
+      }
     }
     else {
       $context = $contexts[$form_state->getValue('machine_name')];
