@@ -2,7 +2,7 @@
 
 namespace Drupal\ctools\Plugin\Deriver;
 
-use Drupal\Core\Plugin\Context\ContextDefinition;
+use Drupal\Core\Plugin\Context\EntityContextDefinition;
 
 /**
  * Deriver that creates a condition for each entity type with bundles.
@@ -17,8 +17,8 @@ class EntityBundle extends EntityDeriverBase {
       if ($entity_type->hasKey('bundle')) {
         $this->derivatives[$entity_type_id] = $base_plugin_definition;
         $this->derivatives[$entity_type_id]['label'] = $this->getEntityBundleLabel($entity_type);
-        $this->derivatives[$entity_type_id]['context'] = [
-          "$entity_type_id" => new ContextDefinition('entity:' . $entity_type_id),
+        $this->derivatives[$entity_type_id]['context_definitions'] = [
+          "$entity_type_id" => new EntityContextDefinition('entity:' . $entity_type_id),
         ];
       }
     }
