@@ -2,7 +2,7 @@
 
 namespace Drupal\ctools\Plugin\Deriver;
 
-use Drupal\Core\Plugin\Context\ContextDefinition;
+use Drupal\Core\Plugin\Context\EntityContextDefinition;
 
 /**
  * Provides entity view block definitions for each entity type.
@@ -17,8 +17,8 @@ class EntityViewDeriver extends EntityDeriverBase {
       if ($entity_type->hasViewBuilderClass()) {
         $this->derivatives[$entity_type_id] = $base_plugin_definition;
         $this->derivatives[$entity_type_id]['admin_label'] = $this->t('Entity view (@label)', ['@label' => $entity_type->getLabel()]);
-        $this->derivatives[$entity_type_id]['context'] = [
-          'entity' => new ContextDefinition('entity:' . $entity_type_id),
+        $this->derivatives[$entity_type_id]['context_definitions'] = [
+          'entity' => new EntityContextDefinition('entity:' . $entity_type_id),
         ];
       }
     }
