@@ -3,7 +3,7 @@
 namespace Drupal\ctools\Plugin\Relationship;
 
 use Drupal\Core\Plugin\Context\Context;
-use Drupal\Core\Plugin\Context\ContextDefinition;
+use Drupal\Core\Plugin\Context\EntityContextDefinition;
 
 /**
  * @Relationship(
@@ -20,7 +20,7 @@ class TypedDataEntityRelationship extends TypedDataRelationship {
     $plugin_definition = $this->getPluginDefinition();
 
     $entity_type = $this->getData($this->getContext('base'))->getDataDefinition()->getSetting('target_type');
-    $context_definition = new ContextDefinition("entity:$entity_type", $plugin_definition['label']);
+    $context_definition = new EntityContextDefinition("entity:$entity_type", $plugin_definition['label']);
     $context_value = NULL;
 
     // If the 'base' context has a value, then get the property value to put on
