@@ -5,8 +5,11 @@ namespace Drupal\ctools\Wizard;
 use Drupal\Core\Form\FormBuilderInterface;
 use Drupal\Core\Form\FormState;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Drupal\Core\Render\Renderer;
+use Drupal\Core\Render\RendererInterface;
 
+/**
+ * The wizard factory.
+ */
 class WizardFactory implements WizardFactoryInterface {
 
   /**
@@ -26,7 +29,7 @@ class WizardFactory implements WizardFactoryInterface {
   /**
    * The object renderer.
    *
-   * @var \Drupal\Core\Render\Renderer
+   * @var \Drupal\Core\Render\RendererInterface
    */
   protected $renderer;
 
@@ -37,10 +40,10 @@ class WizardFactory implements WizardFactoryInterface {
    *   The form builder.
    * @param \Symfony\Component\EventDispatcher\EventDispatcherInterface $event_dispatcher
    *   The event dispatcher.
-   * @param \Drupal\Core\Render\Renderer $renderer
+   * @param \Drupal\Core\Render\RendererInterface $renderer
    *   The object renderer.
    */
-  public function __construct(FormBuilderInterface $form_builder, EventDispatcherInterface $event_dispatcher, Renderer $renderer) {
+  public function __construct(FormBuilderInterface $form_builder, EventDispatcherInterface $event_dispatcher, RendererInterface $renderer) {
     $this->builder = $form_builder;
     $this->dispatcher = $event_dispatcher;
     $this->renderer = $renderer;
@@ -68,12 +71,15 @@ class WizardFactory implements WizardFactoryInterface {
   }
 
   /**
+   * Create form wizard.
+   *
    * @param string $class
    *   A class name implementing FormWizardInterface.
    * @param array $parameters
    *   The array of parameters specific to this wizard.
    *
    * @return \Drupal\ctools\Wizard\FormWizardInterface
+   *   Return form Wizard.
    */
   public function createWizard($class, array $parameters) {
     $arguments = [];
@@ -100,8 +106,10 @@ class WizardFactory implements WizardFactoryInterface {
    * @param array $parameters
    *   The array of parameters specific to this wizard.
    * @param bool $ajax
+   *   Is ajax or not.
    *
    * @return \Drupal\Core\Form\FormState
+   *   Return the form state.
    */
   public function getFormState(FormWizardInterface $wizard, array $parameters, $ajax = FALSE) {
     $form_state = new FormState();
