@@ -3,7 +3,7 @@
 namespace Drupal\Tests\ctools\Kernel;
 
 use Drupal\Core\Plugin\Context\Context;
-use Drupal\Core\Plugin\Context\ContextDefinition;
+use Drupal\Core\Plugin\Context\EntityContextDefinition;
 
 /**
  * @coversDefaultClass \Drupal\ctools\Plugin\RelationshipManager
@@ -22,7 +22,7 @@ class RelationshipManagerTest extends RelationshipsTestBase {
         1 => "foo"
       ]
     ];
-    $this->assertSame($expected, $definitions['typed_data_relationship:entity:node:body']['context']['base']->getConstraints());
+    $this->assertSame($expected, $definitions['typed_data_relationship:entity:node:body']['context_definitions']['base']->getConstraints());
 
     // Check that typed data primitive labels are formatted properly.
     $this->assertSame('Body from Page and Foo', (string) $definitions['typed_data_relationship:entity:node:body']['label']);
@@ -38,21 +38,21 @@ class RelationshipManagerTest extends RelationshipsTestBase {
    * @covers ::getDefinitionsForContexts
    */
   public function testRelationshipPluginAvailability() {
-    $context_definition = new ContextDefinition('entity:node');
+    $context_definition = new EntityContextDefinition('entity:node');
     $contexts = [
       'node' => new Context($context_definition, $this->entities['node1']),
     ];
     $definitions = $this->relationshipManager->getDefinitionsForContexts($contexts);
     //$this->assertTrue(isset($definitions['typed_data_relationship:entity:node:body']));
 
-    $context_definition = new ContextDefinition('entity:node');
+    $context_definition = new EntityContextDefinition('entity:node');
     $contexts = [
       'node' => new Context($context_definition, $this->entities['node2']),
     ];
     $definitions = $this->relationshipManager->getDefinitionsForContexts($contexts);
     $this->assertFalse(isset($definitions['typed_data_relationship:entity:node:body']));
 
-    $context_definition = new ContextDefinition('entity:node');
+    $context_definition = new EntityContextDefinition('entity:node');
     $contexts = [
       'node' => new Context($context_definition, $this->entities['node3']),
     ];
